Refactor files.rb to file-per-method
